API: move workflow model lookups to MCPServer
We want MCPServer to be the only component interacting with workflow
models. In this commit, MCPServer learns two new functions in order to
approve transfers and reingests. The Dashboard is updated accordingly to
invoke these functions and have the old code removed. With this, the Dashboard
API stops using workflow models.

Our pseudo-RPC interface based on Gearman has been updated so the server
can return error messages to the client. Dashboard's RPC client has been
updated to identify these errors and raise a `RPCServerError` exception.

This is connected to #1101.